HUGE EDIT: introduced axios for http requests removed default-pager, string-to-stream, request-promise in refactoring effort main menu gets its own file now index.js looks a lot cleaner reader.js had an overhaul - used child-process to spawn paged content output for articles reader.js used module.parent concept to recursively return to main menu once article ends
