+++ lib/newssrc/editorials.js
@@ -1,19 +1,14 @@
 'use strict';
 const cheerio = require('cheerio');
-const rp = require('request-promise');
-
-function requestContent(url) {
-  return rp(url);
-}
+const https = require('https');
+const axios = require('axios');
+const agent = new https.Agent({  
+  rejectUnauthorized: false
+});
 
 function edsHindu(url, cb) {
-  const options = {
-    url: url,
-    rejectUnauthorized: false
-  };
-
-  requestContent(options).then(function(body) {
-    const $ = cheerio.load(body);
+  axios.get(url, { httpsAgent: agent }).then((response)=> {
+    const $ = cheerio.load(response.data);
     const articles = [];
     $('.100_4x_2EditorialStories')
       .find('h2')
@@ -32,13 +27,8 @@ function edsHindu(url, cb) {
 }
 
 function edsHinduLead(url, cb) {
-  const options = {
-    url: url,
-    rejectUnauthorized: false
-  };
-
-  requestContent(options).then(function(body) {
-    const $ = cheerio.load(body);
+  axios.get(url, { httpsAgent: agent }).then((response)=> {
+    const $ = cheerio.load(response.data);
     const articles = [];
     $('div.story-card-news')
       .find('h3')
@@ -57,13 +47,8 @@ function edsHinduLead(url, cb) {
 }
 
 function edsET(url, cb) {
-  const options = {
-    url: url,
-    rejectUnauthorized: false
-  };
-
-  requestContent(options).then(function(body) {
-    const $ = cheerio.load(body);
+  axios.get(url, { httpsAgent: agent }).then((response)=> {
+    const $ = cheerio.load(response.data);
     const articles = [];
     $('div.content')
       .find('a')
@@ -79,13 +64,8 @@ function edsET(url, cb) {
 }
 
 function edsHBL(url, cb) {
-  const options = {
-    url: url,
-    rejectUnauthorized: false
-  };
-
-  requestContent(options).then(function(body) {
-    const $ = cheerio.load(body);
+  axios.get(url, { httpsAgent: agent }).then((response)=> {
+    const $ = cheerio.load(response.data);
     const articles = [];
     $('article.editorial')
       .find('a:first-child')
@@ -102,13 +82,8 @@ function edsHBL(url, cb) {
 
 //Same function works for both IE editorials and IE columnists
 function edsIndExp(url, cb) {
-  const options = {
-    url: url,
-    rejectUnauthorized: false
-  };
-
-  requestContent(options).then(function(body) {
-    const $ = cheerio.load(body);
+  axios.get(url, { httpsAgent: agent }).then((response)=> {
+    const $ = cheerio.load(response.data);
     const articles = [];
     $('div.profile-container')
       .find('h2')
@@ -125,12 +100,8 @@ function edsIndExp(url, cb) {
 }
 
 function edsGuardian(url, cb) {
-  const options = {
-    url: url,
-    rejectUnauthorized: false
-  };
-  requestContent(options).then(function(body) {
-    const $ = cheerio.load(body);
+  axios.get(url, { httpsAgent: agent }).then((response)=> {
+    const $ = cheerio.load(response.data);
     const articles = [];
     $('.fc-item__content').each(function(i, elem) {
       if (i > 3) {
@@ -153,12 +124,8 @@ function edsGuardian(url, cb) {
 }
 
 function edsLivemint(url, cb) {
-  const options = {
-    url: url,
-    rejectUnauthorized: false
-  };
-  requestContent(options).then(function(body) {
-    const $ = cheerio.load(body);
+  axios.get(url, { httpsAgent: agent }).then((response)=> {
+    const $ = cheerio.load(response.data);
     const articles = [];
     $('.linkclick-candidate')
       .find($('.headline'))
