+++ lib/newssrc/magazines.js
@@ -1,19 +1,14 @@
 'use strict';
 const cheerio = require('cheerio');
-const rp = require('request-promise');
-
-function requestContent(url) {
-  return rp(url);
-}
+const https = require('https');
+const axios = require('axios');
+const agent = new https.Agent({  
+  rejectUnauthorized: false
+});
 
 function magzEconomist(url, cb) {
-  const options = {
-    url: url,
-    rejectUnauthorized: false
-  };
-
-  requestContent(options).then(function(body) {
-    const $ = cheerio.load(body);
+  axios.get(url, { httpsAgent: agent }).then((response)=> {
+    const $ = cheerio.load(response.data);
     const articles = [];
     $('.teaser')
       .find('a')
@@ -30,13 +25,8 @@ function magzEconomist(url, cb) {
 }
 
 function magzAtlantic(url, cb) {
-  const options = {
-    url: url,
-    rejectUnauthorized: false
-  };
-
-  requestContent(options).then(function(body) {
-    const $ = cheerio.load(body);
+  axios.get(url, { httpsAgent: agent }).then((response)=> {
+    const $ = cheerio.load(response.data);
     const articles = [];
     $('.article').each(function(i, elem) {
       const linkExtension = $(this)
@@ -58,13 +48,8 @@ function magzAtlantic(url, cb) {
 }
 
 function magzProjSyndicate(url, cb) {
-  const options = {
-    url: url,
-    rejectUnauthorized: false
-  };
-
-  requestContent(options).then(function(body) {
-    const $ = cheerio.load(body);
+  axios.get(url, { httpsAgent: agent }).then((response)=> {
+    const $ = cheerio.load(response.data);
     const articles = [];
     $('.listing__title')
       .find('a:first-child')
@@ -81,13 +66,8 @@ function magzProjSyndicate(url, cb) {
 }
 
 function magzHBR(url, cb) {
-  const options = {
-    url: url,
-    rejectUnauthorized: false
-  };
-
-  requestContent(options).then(function(body) {
-    const $ = cheerio.load(body);
+  axios.get(url, { httpsAgent: agent }).then((response)=> {
+    const $ = cheerio.load(response.data);
     const articles = [];
     $('h3.hed')
       .find('a:first-child')
@@ -104,13 +84,8 @@ function magzHBR(url, cb) {
 }
 
 function magzFrontline(url, cb) {
-  const options = {
-    url: url,
-    rejectUnauthorized: false
-  };
-
-  requestContent(options).then(function(body) {
-    const $ = cheerio.load(body);
+  axios.get(url, { httpsAgent: agent }).then((response)=> {
+    const $ = cheerio.load(response.data);
     const articles = [];
     $('h2.bigtitle, h2.title')
       .find('a')
@@ -129,13 +104,8 @@ function magzFrontline(url, cb) {
 }
 
 function magzNewYorker(url, cb) {
-  const options = {
-    url: url,
-    rejectUnauthorized: false
-  };
-
-  requestContent(options).then(function(body) {
-    const $ = cheerio.load(body);
+  axios.get(url, { httpsAgent: agent }).then((response)=> {
+    const $ = cheerio.load(response.data);
     const articles = [];
     $('[class^="River__riverItemContent"]')
       .find('a')
@@ -153,13 +123,8 @@ function magzNewYorker(url, cb) {
 }
 
 function magzLongReads(url, cb) {
-  const options = {
-    url: url,
-    rejectUnauthorized: false
-  };
-
-  requestContent(options).then(function(body) {
-    const $ = cheerio.load(body);
+  axios.get(url, { httpsAgent: agent }).then((response)=> {
+    const $ = cheerio.load(response.data);
     const articles = [];
     $('h1.entry-title')
       .find('a')
@@ -175,13 +140,8 @@ function magzLongReads(url, cb) {
 }
 
 function magzHinduExplains(url, cb) {
-  const options = {
-    url: url,
-    rejectUnauthorized: false
-  };
-
-  requestContent(options).then(function(body) {
-    const $ = cheerio.load(body);
+  axios.get(url, { httpsAgent: agent }).then((response)=> {
+    const $ = cheerio.load(response.data);
     const articles = [];
     $('.story-card-news')
       .find('h3')
@@ -202,13 +162,8 @@ function magzHinduExplains(url, cb) {
 }
 
 function magzOutlook(url, cb) {
-  const options = {
-    url: url,
-    rejectUnauthorized: false
-  };
-
-  requestContent(options).then(function(body) {
-    const $ = cheerio.load(body);
+  axios.get(url, { httpsAgent: agent }).then((response)=> {
+    const $ = cheerio.load(response.data);
     const articles = [];
     $('div.category_story_listing8')
       .find('h2, h5')
