+++ lib/newssrc/science.js
@@ -1,19 +1,14 @@
 'use strict';
 const cheerio = require('cheerio');
-const rp = require('request-promise');
-
-function requestContent(url) {
-  return rp(url);
-}
+const https = require('https');
+const axios = require('axios');
+const agent = new https.Agent({  
+  rejectUnauthorized: false
+});
 
 function sciSciAmerican(url, cb) {
-  const options = {
-    url: url,
-    rejectUnauthorized: false
-  };
-
-  requestContent(options).then(function(body) {
-    const $ = cheerio.load(body);
+  axios.get(url, { httpsAgent: agent }).then((response)=> {
+    const $ = cheerio.load(response.data);
     const articles = [];
     $('h2.t_listing-title')
       .find('a')
@@ -29,13 +24,8 @@ function sciSciAmerican(url, cb) {
 }
 
 function sciPopScience(url, cb) {
-  const options = {
-    url: url,
-    rejectUnauthorized: false
-  };
-
-  requestContent(options).then(function(body) {
-    const $ = cheerio.load(body);
+  axios.get(url, { httpsAgent: agent }).then((response)=> {
+    const $ = cheerio.load(response.data);
     const articles = [];
     $('.headline')
       .find('a')
