+++ lib/newssrc/sports.js
@@ -1,19 +1,14 @@
 'use strict';
 const cheerio = require('cheerio');
-const rp = require('request-promise');
-
-function requestContent(url) {
-  return rp(url);
-}
+const https = require('https');
+const axios = require('axios');
+const agent = new https.Agent({  
+  rejectUnauthorized: false
+});
 
 function sportsESPNF1(url, cb) {
-  const options = {
-    url: url,
-    rejectUnauthorized: false
-  };
-
-  requestContent(options).then(function(body) {
-    const $ = cheerio.load(body);
+  axios.get(url, { httpsAgent: agent }).then((response)=> {
+    const $ = cheerio.load(response.data);
     const articles = [];
     $('.contentItem__content').each(function(i, elem) {
       if (i > 6) {
@@ -35,13 +30,8 @@ function sportsESPNF1(url, cb) {
 }
 
 function sportsAutoSportF1(url, cb) {
-  const options = {
-    url: url,
-    rejectUnauthorized: false
-  };
-
-  requestContent(options).then(function(body) {
-    const $ = cheerio.load(body);
+  axios.get(url, { httpsAgent: agent }).then((response)=> {
+    const $ = cheerio.load(response.data);
     const articles = [];
     $('.newsitem').each(function(i, elem) {
       const linkExtension = $(this)
@@ -62,12 +52,8 @@ function sportsAutoSportF1(url, cb) {
 }
 
 function sportsEPL(url, cb) {
-  const options = {
-    url: url,
-    rejectUnauthorized: false
-  };
-  requestContent(options).then(function(body) {
-    const $ = cheerio.load(body);
+  axios.get(url, { httpsAgent: agent }).then((response)=> {
+    const $ = cheerio.load(response.data);
     const articles = [];
     const articleMarkup = $('.featuredArticle').each(function(i, elem) {
       const linkExtension = $(this)
