+++ lib/newssrc/tech.js
@@ -1,19 +1,14 @@
 'use strict';
 const cheerio = require('cheerio');
-const rp = require('request-promise');
-
-function requestContent(url) {
-  return rp(url);
-}
+const https = require('https');
+const axios = require('axios');
+const agent = new https.Agent({  
+  rejectUnauthorized: false
+});
 
 function techTechCrunch(url, cb) {
-  const options = {
-    url: url,
-    rejectUnauthorized: false
-  };
-
-  requestContent(options).then(function(body) {
-    const $ = cheerio.load(body);
+  axios.get(url, { httpsAgent: agent }).then((response)=> {
+    const $ = cheerio.load(response.data);
     const articles = [];
     $('h2.post-block__title')
       .find('a:first-child')
@@ -32,13 +27,8 @@ function techTechCrunch(url, cb) {
 }
 
 function techTNW(url, cb) {
-  const options = {
-    url: url,
-    rejectUnauthorized: false
-  };
-
-  requestContent(options).then(function(body) {
-    const $ = cheerio.load(body);
+  axios.get(url, { httpsAgent: agent }).then((response)=> {
+    const $ = cheerio.load(response.data);
     const articles = [];
     $('h4.story-title')
       .find('a:first-child')
@@ -57,13 +47,8 @@ function techTNW(url, cb) {
 }
 
 function techReadWrite(url, cb) {
-  const options = {
-    url: url,
-    rejectUnauthorized: false
-  };
-
-  requestContent(options).then(function(body) {
-    const $ = cheerio.load(body);
+  axios.get(url, { httpsAgent: agent }).then((response)=> {
+    const $ = cheerio.load(response.data);
     const articles = [];
     $('h2.entry-title')
       .find('a')
@@ -82,13 +67,8 @@ function techReadWrite(url, cb) {
 }
 
 function techBI(url, cb) {
-  const options = {
-    url: url,
-    rejectUnauthorized: false
-  };
-
-  requestContent(options).then(function(body) {
-    const $ = cheerio.load(body);
+  axios.get(url, { httpsAgent: agent }).then((response)=> {
+    const $ = cheerio.load(response.data);
     const articles = [];
     $('.toplist_heading, .liststories_heading')
       .find('a')
