+++ lib/post.js
@@ -82,7 +82,6 @@ function getStories(modifier, options) {
     if (!articles) {
       return deferred.reject();
     }
-
     deferred.resolve(articles);
   });
   return deferred.promise;
