+++ lib/reader.js
@@ -1,11 +1,14 @@
 'use strict';
 
-const rp = require('request-promise');
 const cheerio = require('cheerio');
 const htmlToText = require('html-to-text');
-const pager = require('default-pager');
-const chalk = require('chalk');
-const str = require('string-to-stream')
+const spawn = require("child_process").spawn;
+const https = require('https');
+const axios = require('axios');
+
+const agent = new https.Agent({  
+  rejectUnauthorized: false
+});
 
 function getTextFromHtml(html) {
   return htmlToText.fromString(html, {
@@ -16,8 +19,14 @@ function getTextFromHtml(html) {
   });
 }
 
-function printArticle(content, url) {
-  str(content).pipe(pager());
+function printArticle(content) {
+  const show_content = spawn(`cat <<< "${content}" | less -r`, { 
+      stdio: 'inherit', 
+      shell: true
+    });
+  show_content.on('exit', (code)=> {
+    module.parent.parent.exports.mainMenu();
+  });
 }
 
 function parseContent(options) {
@@ -161,24 +170,15 @@ function readerMagazines($, website) {
   return articleMarkup;
 }
 
-function requestContent(url) {
-  return rp(url);
-}
 
 function show(options) {
-  const optionsParams = {
-    url: options.url,
-    rejectUnauthorized: false
-  };
-
-  requestContent(optionsParams).then(
-    function(body) {
+  axios.get(options.url, { httpsAgent: agent }).then((response)=> {
       const content = parseContent({
-        html: body,
+        html: response.data,
         modifier: options.modifier,
         website: options.website
       });
-      printArticle(content, options.url);
+      printArticle(content);
     },
     function() {
       console.log('Oops! Something went wrong!');
