+++ package.json
@@ -17,6 +17,7 @@
   },
   "homepage": "https://github.com/codePrady/newscast-cli#readme",
   "dependencies": {
+    "axios": "^0.19.2",
     "chalk": "^2.4.2",
     "cheerio": "^1.0.0-rc.3",
     "default-pager": "^1.1.0",
